requirements.txt
requests
prompt_toolkit
python-dotenv